chore: disable analytics for checkly
